pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pipeline_check"
version = "0.1.0"
description = "CI/CD Security Posture Scanner — scores AWS, Terraform, GitHub Actions, GitLab CI, Azure DevOps, Bitbucket Pipelines, and Jenkins against OWASP Top 10 CI/CD Risks"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Daniel Martin" },
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Security",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
]
dependencies = [
    "boto3>=1.34.0",
    "click>=8.1.0",
    "PyYAML>=6.0",
    "rich>=13.0.0",
]

[project.scripts]
pipeline_check = "pipeline_check.cli:scan"

[project.urls]
Homepage = "https://github.com/dmartin/pipeline-check"

[tool.setuptools.packages.find]
include = ["pipeline_check*"]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
]

[tool.ruff.lint.per-file-ignores]
"pipeline_check/core/html_reporter.py" = ["E501"]  # inline CSS/JS templates
"pipeline_check/core/checks/base.py" = ["E402"]    # deferred re import after dataclass
"tests/**" = ["E501"]                               # test strings often exceed line length

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
